% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_lm_intercept_label.R
\name{compute_group_interceptlabel}
\alias{compute_group_interceptlabel}
\title{compute a dataframe for plotting intercept coordinates label}
\usage{
compute_group_interceptlabel(data, scales, digits = 3)
}
\arguments{
\item{data}{a data frame with columns x and y}

\item{digits}{a positive integer}
}
\value{
a data frame with intercept coordinates of x and y and coordinate label
}
\description{
compute a dataframe for plotting intercept coordinates label
}
\examples{
library(dplyr)
library(magrittr)
cars \%>\%
rename(x = speed) \%>\%
rename(y = dist) \%>\%
compute_group_interceptlabel()

}
