% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_text_leverage.R
\name{compute_group_leverage_distance}
\alias{compute_group_leverage_distance}
\title{A function to find leverage distance of a point}
\usage{
compute_group_leverage_distance(data, scales, digits = 3)
}
\arguments{
\item{data}{the data set to be used}

\item{scales}{}
}
\value{
Leverage distance of a certain point
}
\description{
A function to find leverage distance of a point
}
\examples{
library(dplyr)
library(magrittr)
cars \%>\%
mutate(x = speed, y = dist) \%>\%
select(x,y) \%>\%
compute_group_leverage_distance(digits = 2)

}
