% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_point_high_leverage.R
\name{compute_group_leverage_points}
\alias{compute_group_leverage_points}
\title{Function to find Points with High leverage Distances}
\usage{
compute_group_leverage_points(data, scales)
}
\arguments{
\item{data}{the data set to be used}

\item{scales}{}
}
\value{
the points with high leverage values
}
\description{
Function to find Points with High leverage Distances
}
\examples{
library(dplyr)
library(magrittr)
cars\%>\%
mutate(x = speed, y = dist) \%>\%
select(x,y) \%>\%
compute_group_leverage_points()
}
