% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_lm_conf_int.R
\name{geom_lm_conf_int_segments}
\alias{geom_lm_conf_int_segments}
\title{Drawing prediction interval for OLS linear model as segments}
\usage{
geom_lm_conf_int_segments(mapping = NULL, data = NULL,
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{...}{}
}
\description{
Drawing prediction interval for OLS linear model as segments
}
\examples{
library(ggplot2)
ggplot(cars) + aes(x = speed, y = dist) +
geom_point() + geom_lm() + geom_lm_conf_int_segments(num_breaks = 100)
}
