% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_lm_fitted.R
\name{geom_lm_fitted}
\alias{geom_lm_fitted}
\title{Fitted values along ols line, points}
\usage{
geom_lm_fitted(mapping = NULL, data = NULL, position = "identity",
  na.rm = FALSE, show.legend = NA, inherit.aes = TRUE, ...)
}
\arguments{
\item{...}{}
}
\description{
Fitted values along ols line, points
}
\examples{
library(ggplot2)
ggplot(cars) + aes(x = speed, y = dist) +
geom_point() + geom_lm() + geom_lm_fitted(color = "blue")
}
