% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_lm_formula.R
\name{geom_lm_formula}
\alias{geom_lm_formula}
\title{Write formula for ols linear model}
\usage{
geom_lm_formula(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Write formula for ols linear model
}
\examples{
library(ggplot2)
ggplot(cars) + aes(x = speed, y = dist) +
geom_point() + geom_lm() + geom_lm_formula()
}
