% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_lm_pooled_indicator.R
\name{geom_lm_indicator}
\alias{geom_lm_indicator}
\title{Drawing prediction interval for OLS linear model}
\usage{
geom_lm_indicator(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = T,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Drawing prediction interval for OLS linear model
}
\examples{
library(ggplot2)
library(dplyr)
library(magrittr)

starwars \%>\%
   filter(mass < 1000) \%>\%
   tidyr::drop_na(mass, height, sex) ->
 my_data

 my_data \%>\%
   lm(mass ~ height + sex, data = .) ->
 model

data.frame(x = my_data$height,
y = model$fitted.values,
group = my_data$sex) \%>\%
ggplot() +
aes(x = x, y = y, group = group) +
geom_point() +
geom_line()

starwars \%>\%
filter(mass < 1000) \%>\%
mutate(sex_numeric = sex \%>\% as.factor() \%>\% as.numeric())  \%>\%
ggplot() +
aes(x = height, y = mass, indicator_as_numeric = sex_numeric) +
geom_point(aes(color = sex)) + geom_lm_indicator(linetype = "dashed")
}
