% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_lm_intercept.R
\name{geom_lm_intercept}
\alias{geom_lm_intercept}
\title{Intercept}
\usage{
geom_lm_intercept(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Intercept
}
\examples{
library(ggplot2)
ggplot(cars) + aes(x = speed, y = dist) +
geom_point() + geom_lm() + geom_lm_intercept(color = "blue")
}
