% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_lm_intercept_label.R
\name{geom_lm_intercept_label}
\alias{geom_lm_intercept_label}
\title{Label ols linear model intercept}
\usage{
geom_lm_intercept_label(mapping = NULL, data = NULL,
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{...}{}
}
\description{
Label ols linear model intercept
}
\examples{
library(ggplot2)
ggplot(cars) + aes(x = speed, y = dist) +
geom_point() + geom_lm() + geom_lm_intercept(color = "blue") +
geom_lm_intercept_label(hjust = 0)
}
