% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_lm_residuals.R
\name{geom_lm_residuals}
\alias{geom_lm_residuals}
\title{Drawing residuals from ols linear model}
\usage{
geom_lm_residuals(mapping = NULL, data = NULL, position = "identity",
  na.rm = FALSE, show.legend = NA, inherit.aes = TRUE, ...)
}
\arguments{
\item{...}{}
}
\description{
Drawing residuals from ols linear model
}
\examples{
library(ggplot2)
ggplot(cars) + aes(x = speed, y = dist) +
geom_point() + geom_lm() + geom_lm_residuals(color = "darkred")
}
