% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_normal_dist.R
\name{geom_normal_dist}
\alias{geom_normal_dist}
\title{Fits normal distribution based on standard deviation}
\usage{
geom_normal_dist(mapping = NULL, data = NULL, position = "identity",
  na.rm = FALSE, show.legend = NA, inherit.aes = TRUE,
  outline.type = "upper", ...)
}
\arguments{
\item{...}{}
}
\description{
Fits normal distribution based on standard deviation
}
\examples{
library(ggplot2)
ggplot(faithful, aes(x = waiting)) +
  geom_rug() +
  geom_histogram(aes(y = ..density..)) +
  geom_normal_dist(fill = "magenta") +
  facet_wrap(facets = vars( eruptions > 3 ), ncol = 1)
}
