% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_normal_dist.R
\name{geom_normal_dist_zlines}
\alias{geom_normal_dist_zlines}
\title{Fits normal distribution based on standard deviation}
\usage{
geom_normal_dist_zlines(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  outline.type = "upper",
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Fits normal distribution based on standard deviation
}
\examples{
library(ggplot2)
ggplot(faithful, aes(x = waiting)) +
  geom_rug() +
  geom_histogram(aes(y = ..density..)) +
  geom_normal_dist(fill = "magenta") +
  geom_normal_dist_zlines() +
  facet_wrap(facets = vars( eruptions > 3 ), ncol = 1) +
  geom_normal_dist(q_min = .99)

ggplot(faithful, aes(x = waiting)) +
  geom_density() +
  geom_normal_dist(q_min = .05, q_max = 1-.05)
}
