% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_point_high_cooks.R
\name{geom_point_high_cooks}
\alias{geom_point_high_cooks}
\title{Returns a scatter plot with points that are colored if they have a high cooks distance}
\usage{
geom_point_high_cooks(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\value{
a scatter plot with points that are colored if they have a high cooks distance
}
\description{
Returns a scatter plot with points that are colored if they have a high cooks distance
}
\examples{
mycars \%>\%
ggplot(aes(x = x, y = y)) +
geom_point() +
geom_point_high_cooks(color = 'red')
}
