% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_point_high_leverage.R
\name{geom_point_high_leverage}
\alias{geom_point_high_leverage}
\title{Returns a scatter plot with points that are colored if they have a high leverage distance}
\usage{
geom_point_high_leverage(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\value{
a scatter plot with points that are colored if they have a high leverage distance
}
\description{
Returns a scatter plot with points that are colored if they have a high leverage distance
}
\examples{
library(ggplot2)
library(magrittr)
cars \%>\%
ggplot(aes(x = speed, y = dist)) +
geom_point() +
geom_point_high_leverage(color = 'red')
}
