% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_text_cooks.R
\name{geom_text_cooks}
\alias{geom_text_cooks}
\title{Returns a scatter plot with points that are labeled}
\usage{
geom_text_cooks(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\value{
a scatter plot with points that are labeled
}
\description{
Returns a scatter plot with points that are labeled
}
\examples{
cars \%>\%
ggplot(aes(x = speed, y = dist)) +
geom_point() +
geom_lm() +
geom_text_cooks(check_overlap = TRUE,
digits = 2)
}
