% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_x_label.R
\name{geom_x_label}
\alias{geom_x_label}
\title{Labels defined by values of x}
\usage{
geom_x_label(
  mapping = NULL,
  data = NULL,
  ...,
  x,
  na.rm = FALSE,
  show.legend = NA
)
}
\arguments{
\item{show.legend}{}
}
\value{

}
\description{
Labels defined by values of x
}
