% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_x_line.R
\name{geom_x_line}
\alias{geom_x_line}
\title{Lines defined by values of x}
\usage{
geom_x_line(mapping = NULL, data = NULL, ..., x, na.rm = FALSE,
  show.legend = NA)
}
\arguments{
\item{show.legend}{}
}
\description{
Lines defined by values of x
}
\examples{
library(ggplot2)
ggplot(data = cars, mapping = aes(x = speed, y = dist)) +
geom_point() + geom_x_line()
}
