% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_x_mean_label.R
\name{geom_x_mean_label}
\alias{geom_x_mean_label}
\title{Lines defined by mean value of y}
\usage{
geom_x_mean_label(
  mapping = NULL,
  data = NULL,
  ...,
  x,
  na.rm = FALSE,
  show.legend = NA
)
}
\arguments{
\item{show.legend}{}
}
\value{

}
\description{
Lines defined by mean value of y
}
\examples{
library(ggplot2)
ggplot(data = cars, mapping = aes(x = speed, y = dist)) +
geom_point() + geom_x_mean() + geom_x_mean_label()
}
