% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_x_percentile.R
\name{geom_x_percentile}
\alias{geom_x_percentile}
\title{Lines defined by quantile value of x}
\usage{
geom_x_percentile(
  mapping = NULL,
  data = NULL,
  ...,
  x,
  na.rm = FALSE,
  show.legend = NA
)
}
\arguments{
\item{show.legend}{}
}
\value{

}
\description{
Lines defined by quantile value of x
}
\examples{
library(ggplot2)
ggplot(data = cars, mapping = aes(x = speed, y = dist)) +
geom_point() + geom_x_percentile(percentile = c(25,75), linetype = "dashed") +
geom_x_percentile(percentile = 0, linetype = "dotted")
}
