% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_x_quantile.R
\name{geom_x_quantile}
\alias{geom_x_quantile}
\title{Lines defined by quantile value of x}
\usage{
geom_x_quantile(mapping = NULL, data = NULL, ..., x, na.rm = FALSE,
  show.legend = NA)
}
\arguments{
\item{show.legend}{}
}
\description{
Lines defined by quantile value of x
}
\examples{
library(ggplot2)
ggplot(data = cars, mapping = aes(x = speed, y = dist)) +
geom_point() + geom_x_quantile(quantile = c(.25,.75), linetype = "dashed")
}
