% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xvlines.R
\name{geom_xmean}
\alias{geom_xmean}
\title{Title}
\usage{
geom_xmean(mapping = NULL, data = NULL, stat = "xmean",
  position = "identity", ..., arrow = NULL, lineend = "butt",
  linejoin = "round", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE)
}
\arguments{
\item{inherit.aes}{}
}
\description{
Title
}
