% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xvlines.R
\name{geom_xvlines}
\alias{geom_xvlines}
\title{Title}
\usage{
geom_xvlines(mapping = NULL, data = NULL, stat = "xvlines",
  position = "identity", ..., r = 1, n = 50, arrow = NULL,
  lineend = "butt", linejoin = "round", na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE)
}
\arguments{
\item{inherit.aes}{}
}
\description{
Title
}
