% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_xy_means.R
\name{geom_xy_means}
\alias{geom_xy_means}
\title{Place point at mean of x and mean of y}
\usage{
geom_xy_means(mapping = NULL, data = NULL, position = "identity",
  na.rm = FALSE, show.legend = NA, inherit.aes = TRUE, ...)
}
\arguments{
\item{...}{}
}
\description{
Place point at mean of x and mean of y
}
\examples{
library(ggplot2)
ggplot(cars) + aes(x = speed, y = dist) +
geom_point() + geom_xy_means(size = 5, color = "plum4")
}
