% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_segment_y_ymean.R
\name{geom_xy_xymean}
\alias{geom_xy_xymean}
\title{Title}
\usage{
geom_xy_xymean(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
