% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_y_line.R
\name{geom_y_line}
\alias{geom_y_line}
\title{Lines defined by values of y}
\usage{
geom_y_line(mapping = NULL, data = NULL, ..., y, na.rm = FALSE,
  show.legend = NA)
}
\arguments{
\item{show.legend}{}
}
\description{
Lines defined by values of y
}
\examples{
library(ggplot2)
ggplot(data = cars, mapping = aes(x = speed, y = dist)) +
geom_point() + geom_y_line()
}
