% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_y_mean_label.R
\name{geom_y_mean_label}
\alias{geom_y_mean_label}
\title{Lines defined by mean value of y}
\usage{
geom_y_mean_label(mapping = NULL, data = NULL, ..., y, na.rm = FALSE,
  show.legend = NA)
}
\arguments{
\item{show.legend}{}
}
\description{
Lines defined by mean value of y
}
\examples{
library(ggplot2)
ggplot(data = cars, mapping = aes(x = speed, y = dist)) +
geom_point() + geom_y_mean_label() + geom_y_mean()
}
