% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xvyhline.R
\name{geom_yhline}
\alias{geom_yhline}
\title{Draw vertical line at each value of y}
\usage{
geom_yhline(mapping = NULL, data = NULL, position = "identity",
  na.rm = FALSE, show.legend = NA, inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{provide you own mapping, y is required}

\item{data}{provide you own data}

\item{position}{change geom}

\item{na.rm}{remove missing values without warning, default is F}

\item{show.legend}{show legend in plot, default is T}

\item{inherit.aes}{should the geom inherits aesthetics from global, default is true}

\item{...}{other arguments to be passed to the geom, see geom segment}
}
\value{
a ggplot2 layer
}
\description{
Draw vertical line at each value of y
}
\examples{
library(ggplot2)
ggplot(data = cars, mapping = aes(x = speed, y = dist)) + geom_point() + geom_yhline()
}
