% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_normal_dist.R
\name{stamp_normal_dist_zlines}
\alias{stamp_normal_dist_zlines}
\title{Title}
\usage{
stamp_normal_dist_zlines(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  outline.type = "upper",
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
\examples{
library(ggplot2)
library(tidyverse)
ggxmean:::stamp_space() +
  stamp_normal_dist(alpha = .5, height = 1, fill = "magenta") +
  stamp_normal_dist(sd_min = -5, sd_max = -1.96, height = 1) +
  stamp_normal_dist(q_min = 0.25, height = 1) +
  stamp_normal_dist_zlines()

ggxmean:::stamp_space() +
  stamp_normal_dist(sd_min = -1, sd_max = 1, color = "slateblue",
             color = "grey34", outline.type = "full") +
  stamp_normal_dist(sd_min = -2, sd_max = 2, color = "slateblue", outline.type = "full") +
  stamp_normal_dist(sd_min = -3, sd_max = 3, color = "slateblue", outline.type = "full") +
  stamp_normal_dist(sd_min = -4, sd_max = 4, color = "slateblue", outline.type = "full") +
  stamp_normal_dist(sd_min = -5, sd_max = 5, color = "slateblue", outline.type = "full")
}
