% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_t_dist.R
\name{stamp_t_prob}
\alias{stamp_t_prob}
\title{Title}
\usage{
stamp_t_prob(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  outline.type = "upper",
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
\examples{
library(ggplot2)
ggplot(cars, aes(x = dist)) +
  stamp_t_prob(alpha = .5, height = 1, fill = "magenta") +
  stamp_t_prob(x_min = -5, x_max = -1.96, height = 1) +
  stamp_t_prob(x_min = 1.98, x_max = 5, height = 1)

ggplot(cars, aes(x = dist)) +
  stamp_t_prob(x_min = -1, x_max = 1, color = "slateblue",
             color = "grey34", outline.type = "full") +
  stamp_t_prob(x_min = -2, x_max = 2, color = "slateblue", outline.type = "full") +
  stamp_t_prob(x_min = -3, x_max = 3, color = "slateblue", outline.type = "full") +
  stamp_t_prob(x_min = -4, x_max = 4, color = "slateblue", outline.type = "full") +
  stamp_t_prob(x_min = -5, x_max = 5, color = "slateblue", outline.type = "full")
}
