% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stamp_weibull_dist.R
\name{stamp_weibull_dist}
\alias{stamp_weibull_dist}
\title{Title}
\usage{
stamp_weibull_dist(mapping = NULL, data = NULL,
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, outline.type = "upper", ...)

stamp_weibull_dist(mapping = NULL, data = NULL,
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, outline.type = "upper", ...)
}
\arguments{
\item{...}{}

\item{...}{}
}
\description{
Title

Title
}
\examples{
library(ggplot2)
ggplot(cars, aes(x = dist)) +
  stamp_weibull_dist(alpha = .5, height = 1, fill = "magenta") +
  stamp_weibull_dist(sd_min = -5, sd_max = -1.96, height = 1) +
  stamp_weibull_dist(sd_min = 1.98, sd_max = 5, height = 1)

ggplot(cars, aes(x = dist)) +
  stamp_weibull_dist(sd_min = -1, sd_max = 1, color = "slateblue",
             color = "grey34", outline.type = "full") +
  stamp_weibull_dist(sd_min = -2, sd_max = 2, color = "slateblue", outline.type = "full") +
  stamp_weibull_dist(sd_min = -3, sd_max = 3, color = "slateblue", outline.type = "full") +
  stamp_weibull_dist(sd_min = -4, sd_max = 4, color = "slateblue", outline.type = "full") +
  stamp_weibull_dist(sd_min = -5, sd_max = 5, color = "slateblue", outline.type = "full")
library(ggplot2)
ggplot(cars, aes(x = dist)) +
  stamp_weibull_dist(alpha = .5, height = 1, fill = "magenta") +
  stamp_weibull_dist(sd_min = -5, sd_max = -1.96, height = 1) +
  stamp_weibull_dist(sd_min = 1.98, sd_max = 5, height = 1)

ggplot(cars, aes(x = dist)) +
  stamp_weibull_dist(sd_min = -1, sd_max = 1, color = "slateblue",
             color = "grey34", outline.type = "full") +
  stamp_weibull_dist(sd_min = -2, sd_max = 2, color = "slateblue", outline.type = "full") +
  stamp_weibull_dist(sd_min = -3, sd_max = 3, color = "slateblue", outline.type = "full") +
  stamp_weibull_dist(sd_min = -4, sd_max = 4, color = "slateblue", outline.type = "full") +
  stamp_weibull_dist(sd_min = -5, sd_max = 5, color = "slateblue", outline.type = "full")
}
