% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_x_mean.R
\name{geom_x_mean}
\alias{geom_x_mean}
\title{Lines defined by mean value of x}
\usage{
geom_x_mean(mapping = NULL, data = NULL, ..., x, na.rm = FALSE,
  show.legend = NA)
}
\arguments{
\item{show.legend}{}
}
\description{
Lines defined by mean value of x
}
\examples{
library(ggplot2)
ggplot(data = cars, mapping = aes(x = speed, y = dist)) +
geom_point() + geom_x_mean()
}
