% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_y_mean.R
\name{geom_y_mean}
\alias{geom_y_mean}
\title{Place line at mean of y}
\usage{
geom_y_mean(
  mapping = NULL,
  data = NULL,
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Place line at mean of y
}
\examples{
library(ggplot2)
ggplot(cars) + aes(x = speed, y = dist) +
geom_point() + geom_y_mean(size = 1.5) + aes(color = speed > 14)
}
