% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_y_median.R
\name{geom_y_median}
\alias{geom_y_median}
\title{Lines defined by median value of y}
\usage{
geom_y_median(
  mapping = NULL,
  data = NULL,
  ...,
  y,
  na.rm = FALSE,
  show.legend = NA
)
}
\arguments{
\item{show.legend}{}
}
\value{

}
\description{
Lines defined by median value of y
}
\examples{
library(ggplot2)
ggplot(data = cars, mapping = aes(x = speed, y = dist)) +
geom_point() + geom_y_median()
}
